mengatur route dan permission untuk tata_usaha

+++ resources/views/sekolah/index.blade.php
@@ -49,57 +49,61 @@
                             </div>
                             <div class="card-header">
                                 <div class="row">
-                                    <div class="col-sm-8">
-                                        <button type="button" class="btn btn-primary float-left btn-sm" data-toggle="modal"
-                                            data-target="#staticBackdrop">
-                                            Tambah Data Sekolah
-                                        </button>
-                                    </div>
+                                    @if (auth()->user()->role == 'admin')
+                                        <div class="col-sm-8">
+                                            <button type="button" class="btn btn-primary float-left btn-sm"
+                                                data-toggle="modal" data-target="#staticBackdrop">
+                                                Tambah Data Sekolah
+                                            </button>
+                                        </div>
 
-                                    <div class="col-sm-1">
-                                        <button type="button" class="btn btn-primary btn-sm" data-toggle="modal"
-                                            data-target="#importExcel">
-                                            IMPOR EXCEL
-                                        </button>
-                                        <!-- Import Excel -->
-                                        <div class="modal fade" id="importExcel" tabindex="-1" role="dialog"
-                                            aria-labelledby="exampleModalLabel" aria-hidden="true">
-                                            <div class="modal-dialog" role="document">
-                                                <form method="post" action="/sekolah/import_excel"
-                                                    enctype="multipart/form-data">
-                                                    <div class="modal-content">
-                                                        <div class="modal-header">
-                                                            <h5 class="modal-title" id="exampleModalLabel">Import Excel
-                                                            </h5>
-                                                        </div>
-                                                        <div class="modal-body">
-
-                                                            {{ csrf_field() }}
-
-                                                            <label>Pilih file excel</label>
-                                                            <div class="form-group">
-                                                                <input type="file" name="file" required="required">
+                                        <div class="col-sm-1">
+                                            <button type="button" class="btn btn-primary btn-sm" data-toggle="modal"
+                                                data-target="#importExcel">
+                                                IMPOR EXCEL
+                                            </button>
+                                            <!-- Import Excel -->
+                                            <div class="modal fade" id="importExcel" tabindex="-1" role="dialog"
+                                                aria-labelledby="exampleModalLabel" aria-hidden="true">
+                                                <div class="modal-dialog" role="document">
+                                                    <form method="post" action="/sekolah/import_excel"
+                                                        enctype="multipart/form-data">
+                                                        <div class="modal-content">
+                                                            <div class="modal-header">
+                                                                <h5 class="modal-title" id="exampleModalLabel">Import Excel
+                                                                </h5>
                                                             </div>
+                                                            <div class="modal-body">
 
+                                                                {{ csrf_field() }}
+
+                                                                <label>Pilih file excel</label>
+                                                                <div class="form-group">
+                                                                    <input type="file" name="file"
+                                                                        required="required">
+                                                                </div>
+
+                                                            </div>
+                                                            <div class="modal-footer">
+                                                                <button type="button" class="btn btn-secondary"
+                                                                    data-dismiss="modal">Close</button>
+                                                                <button type="submit"
+                                                                    class="btn btn-primary">Import</button>
+                                                            </div>
                                                         </div>
-                                                        <div class="modal-footer">
-                                                            <button type="button" class="btn btn-secondary"
-                                                                data-dismiss="modal">Close</button>
-                                                            <button type="submit" class="btn btn-primary">Import</button>
-                                                        </div>
-                                                    </div>
-                                                </form>
+                                                    </form>
+                                                </div>
                                             </div>
                                         </div>
-                                    </div>
-                                    <div class="col-sm-1">
-                                        <a href="/sekolah/export_excel" class="btn btn-primary float-right btn-sm"
-                                            target="_blank">EXPOR EXCEL</a>
-                                    </div>
-                                    <div class="col-sm-1">
-                                        <a href="/sekolah/export_pdf" class="btn btn-primary float-right btn-sm"
-                                            target="_blank">EXPOR PDF</a>
-                                    </div>
+                                        <div class="col-sm-1">
+                                            <a href="/sekolah/export_excel" class="btn btn-primary float-right btn-sm"
+                                                target="_blank">EXPOR EXCEL</a>
+                                        </div>
+                                        <div class="col-sm-1">
+                                            <a href="/sekolah/export_pdf" class="btn btn-primary float-right btn-sm"
+                                                target="_blank">EXPOR PDF</a>
+                                        </div>
+                                    @endif
                                 </div>
                             </div>
 
@@ -136,12 +140,13 @@
                                                     <a href="/sekolah/{{ $sekolah->id }}/sekolahedit"
                                                         class="btn btn-warning btn-sm">Ubah
                                                     </a>
-                                                    <button href="/sekolah/{{ $sekolah->id }}/sekolahdelete"
-                                                        type="button" class="btn btn-danger btn-sm" data-toggle="modal"
-                                                        data-target="#modal-danger">
-                                                        Hapus
-                                                    </button>
-
+                                                    @if (auth()->user()->role == 'admin')
+                                                        <button href="/sekolah/{{ $sekolah->id }}/sekolahdelete"
+                                                            type="button" class="btn btn-danger btn-sm" data-toggle="modal"
+                                                            data-target="#modal-danger">
+                                                            Hapus
+                                                        </button>
+                                                    @endif
                                                 </td>
                                             </tr>
                                         @endforeach
